+++ src/util.js
@@ -12,20 +12,71 @@ function getChecksum(callName, queryParams, sharedSecret, hashMethod = 'sha1') {
 }
 
 function constructUrl(options, action, params) {
-  params.checksum = getChecksum(action, params, options.salt, options.hashMethod)
-  return `${options.host}/api/${action}?${querystring.encode(params)}`
+  const queryParams = { ...params }
+  queryParams.checksum = getChecksum(
+    action,
+    queryParams,
+    options.salt,
+    options.hashMethod
+  )
+
+  return `${options.host}/api/${action}?${querystring.encode(queryParams)}`
 }
 
-function httpClient(url) {
-  return axios(url, {
-    headers: { Accept: 'text/xml, application/json, text/plain, */*' },
-  })
-    .then((response) => {
-      return response.data
-    })
-    .then(function (xml) {
-      return parseXml(xml)
-    })
+function constructRequest(options, action, params, requestOptions = {}) {
+  const url = constructUrl(options, action, params)
+
+  return {
+    method: requestOptions.method || 'get',
+    url,
+    headers: {
+      Accept: 'text/xml, application/json, text/plain, */*',
+      ...(requestOptions.headers || {}),
+    },
+    data: requestOptions.data,
+  }
+}
+
+function parseResponse(response) {
+  const responseData = response && response.data
+  const contentType =
+    (response && response.headers && response.headers['content-type']) || ''
+
+  if (responseData && typeof responseData === 'object') {
+    return responseData.response || responseData
+  }
+
+  if (typeof responseData !== 'string') {
+    return responseData
+  }
+
+  if (/json/i.test(contentType)) {
+    const json = JSON.parse(responseData)
+    return json.response || json
+  }
+
+  return parseXml(responseData)
+}
+
+function httpClient(urlOrRequest, requester = axios) {
+  const request =
+    typeof urlOrRequest === 'string'
+      ? {
+          url: urlOrRequest,
+          method: 'get',
+          headers: { Accept: 'text/xml, application/json, text/plain, */*' },
+        }
+      : {
+          method: 'get',
+          headers: { Accept: 'text/xml, application/json, text/plain, */*' },
+          ...urlOrRequest,
+          headers: {
+            Accept: 'text/xml, application/json, text/plain, */*',
+            ...(urlOrRequest && urlOrRequest.headers ? urlOrRequest.headers : {}),
+          },
+        }
+
+  return requester(request).then(parseResponse)
 }
 
 function normalizeUrl(url) {
@@ -39,7 +90,7 @@ function getPathname(url, host) {
 function parseXml(xml) {
   const json = parser.parse(xml).response
 
-  if(json.meetings) {
+  if (json.meetings) {
     let meetings = json.meetings ? json.meetings.meeting : []
     meetings = Array.isArray(meetings) ? meetings : [meetings]
     json.meetings = meetings
@@ -55,7 +106,9 @@ function parseXml(xml) {
 module.exports = {
   httpClient,
   constructUrl,
+  constructRequest,
   normalizeUrl,
   getPathname,
   parseXml,
+  parseResponse,
 }